map response

+++ Decode.PackagesApi.Distributor/Common/Models/Price.cs
@@ -2,7 +2,7 @@
 {
     public class Price
     {
-        public int? Total { get; set; }
+        public decimal? Total { get; set; }
         public string Currency { get; set; }
 
     }
